some new packages with new functionality. Changed to using BigDecimal rather than double because double doesn't retain precision whereas BigDecimal does. Also added new interface for measures that need to return string values. Added a parent interface that the BigDecimal and String specific measure interfaces extend so that the classmap will be able to behave as intended. Added in UIServices that would be the gateway between UI and Backend. Added TableUtilities that are the structures that will store the data of the UI table. The DataSet class is the structure that will be used to transfer data around.

+++ src/main/java/Main.java
@@ -1,5 +1,6 @@
 import BackEndUtilities.*;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
@@ -9,9 +10,9 @@ public class Main {
         DataSet ds = new DataSet();
         ds.addData(Arrays.asList("x*12", "y/3", "y*12", "x/3", "12", "19", "72"));
         ds.addVariables(Arrays.asList("x=3", "y=12"));
-        double output = (double) FunctionCaller.measureRunner("standard deviation", ds);
+        BigDecimal output = (BigDecimal) FunctionCaller.measureRunner("standard deviation", ds);
 
-        List<Double> value = Expressions.eval(ds);
+        List<BigDecimal> value = Expressions.eval(ds);
         value.sort(Comparator.naturalOrder());
         value.forEach(System.out::println);
         System.out.println("standard deviation: " + output);
